+++ os_specific.cpp
@@ -4,7 +4,7 @@
 #include <exception>
 
 #include "os_specific.h"
-#include "io.h"
+#include "io_api.h"
 
 // OS specific preprocessor definitions.
 //
